+++ src/cuda/gemm_tensorcores/gemm_optimized/sgemm_nn_64_16_16_16_4.h
@@ -14,7 +14,7 @@ void sgemm(cudaStream_t stream, float *C, const float *A, const float *B,
 
 
 
-void sgemm_dmr(cudaStream_t stream, double *C, double *C_inc, const double *A,
+void sgemm_dmr(cudaStream_t stream, double *C, float *C_inc, const double *A,
 		const double *B, int32_t m, int32_t n, int32_t k, int32_t lda,
 		int32_t ldb, int32_t ldc, double alpha, double beta);
